# Simplify decide_worker