@gjoseph92 2022-08-31
> Basically, we'd slow down embarrassingly parallel submissions, e.g. `client.map(inc, range(1000))` by ~100ms if scheduled on a 1k workers cluster

A plot of your data table:

![image](https://user-images.githubusercontent.com/3309802/187483319-9df117fc-c815-4c6f-96e5-0c347b1a38a7.png)

1k workers isn't unreasonable. Even 10k happens sometimes. Another thing to note is that in really large clusters like that, you'd have to have a ton of root tasks for them to exceed the `total_nthreads * 2` limit. So perhaps with larger clusters, it would actually be more likely to end up in this fastpath, which is also the case where the fastpath would make the most difference?

> This code path is in a real world scenario actually pretty difficult to hit since we introduced the root task logic above

I think this is the more important point. `n_tasks < total_nthreads * 2` puts an upper bound on how bad performance can be (with tons of workers, that could still be a high bound though, see above). The other way a task could avoid the root task check is if the TaskGroup has >5 dependencies, in which case it wouldn't get to use the zero-deps fastpath anyway.

The main thing that worries me is that TaskGroups are a very brittle way of inferring graph structure. All you have to do is submit a bunch of tasks with UUIDs as keys, and they'll bypass the root task logic since they'll all belong to different TaskGroups.

That again points to the importance of determining root-ish-ness from the graph itself, not task names: https://github.com/dask/distributed/issues/6922.

------

Basically I'd be a little hesitant to just remove this logic. Instead, I think we should first work on the definition of `is_rootish` so that all tasks with 0 deps fall within it, regardless of TaskGroup/cluster size. Then, we could even add a similar fastpath to `decide_worker_rootish_*`. I'm imagining something like:
```python
if len(self.idle) > 100:
    ws = random.choice(self.idle.values())
else:
    ws = min(self.idle.values(), key=lambda ws: len(ws.processing) / ws.nthreads)
```