@fjetter 2022-08-31
> A plot of your data table:

I know this plot looks dramatic but it is not surprising. What we do here, sorting, scales with `W Log(W)` where W is the number of workers. So, yes, for 10k we're likely already in a ~1-2s range for a thousand tasks.

My point is basically that when we normalize these numbers to overhead per task this doesn't feel to be too dramatic anymore and I am doubtful that real world workloads would really notice. after all, if you are running on 10k workers I expect you're processing *a lot* of data. Would 1s overhead for the initial dispatch overhead?


Basically, I think having fewer branches in the decide_worker logic would make reasoning about what happens much easier and that might be worth the overhead