@fjetter 2022-08-31
If we dropped this fast path we could also demote the workers dictionary to a plain `dict`. It is currently a `sortedcontainers.SortedDict`.
The measurements significantly improve, specifically in the high worker range. We still have the non-linear scaling but as I mentioned above, I don't think we should really care since the absolute values are really small